+++ .vscode/settings.json
@@ -1,3 +1,4 @@
 {
-    "nuxt.isNuxtApp": false
+  "nuxt.isNuxtApp": false,
+  "cSpell.words": ["Hackathon"]
 }
