--- apps/client/src/app/hackathon/_components/index.ts
@@ -1,4 +1,0 @@
-import { Record } from './Record'
-import { Column } from './Record/Column'
-import { Body } from './Record/Body'
-export { Record, Column, Body }
