+++ apps/client/src/features/HackathonRecord/components/Record/index.ts
@@ -1,0 +1,4 @@
+import { Body } from '@client/features/HackathonRecord/components/Record/Body/Body'
+import { Column } from '@client/features/HackathonRecord/components/Record/Column/Column'
+
+export { Body, Column }
