+++ apps/client/src/features/HackathonRecord/components/index.ts
@@ -1,0 +1,3 @@
+import { Record } from './Record/Record'
+
+export { Record }
